Refactor DataViews metadata normalization
