+++ packages/ui/src/runtime/dataviews/metadata/index.ts
@@ -1,0 +1,7 @@
+export {
+	DATA_VIEWS_METADATA_INVALID,
+	type DataViewMetadataIssue,
+	type MetadataNormalizationResult,
+	type ResourceDataViewMetadata,
+} from './types';
+export { normalizeResourceDataViewMetadata } from './normalize';
